feat/refact repository test
